move markdown viewer out of its own dir
